Update quickstart to make the second half more concrete (#45)
The second half of the quickstart was a bit hand wavy, just suggesting
that the user "in a second process (such as another tab) put the
following code". This commit uses the @automerge/create-repo-node-app
helper (yet to be merged:
https://github.com/automerge/automerge-repo/pull/224) to guide the
reader through creating a node app to use as the second process.